Moved extra DSP functions to utility/dspinst.h
