+++ effect_midside_encode.h
@@ -25,29 +25,7 @@
 
 #include "AudioStream.h"
 
-//#include "utility/dspinst.h"
-// additional dspinst:
-// computes out = (((a[31:16]+b[31:16])/2) <<16) | ((a[15:0]+b[15:0])/2)
-static inline int32_t signed_halving_add_16_and_16(int32_t a, int32_t b) __attribute__((always_inline, unused));
-static inline int32_t signed_halving_add_16_and_16(int32_t a, int32_t b)
-{
-	int32_t out;
-	asm volatile("shadd16 %0, %1, %2" : "=r" (out) : "r" (a), "r" (b));
-	return out;
-}
-
-// additional dspinst:
-// computes out = (((a[31:16]-b[31:16])/2) <<16) | ((a[15:0]-b[15:0])/2)
-static inline int32_t signed_halving_subtract_16_and_16(int32_t a, int32_t b) __attribute__((always_inline, unused));
-static inline int32_t signed_halving_subtract_16_and_16(int32_t a, int32_t b)
-{
-	int32_t out;
-	asm volatile("shsub16 %0, %1, %2" : "=r" (out) : "r" (a), "r" (b));
-	return out;
-}
-
-
-
+#include "utility/dspinst.h"
 
 /**
  * This object performs the encoding of a stereo signal to its mid/side components.
