+++ utility/dspinst.h
@@ -38,6 +38,17 @@ static inline int32_t signed_saturate_rshift(int32_t val, int bits, int rshift)
 	return out;
 }
 
+// computes limit(val, 2**bits)
+static inline int16_t saturate16(int32_t val) __attribute__((always_inline, unused));
+static inline int16_t saturate16(int32_t val)
+{
+	int16_t out;
+	int32_t tmp;
+	asm volatile("ssat %0, %1, %2" : "=r" (tmp) : "I" (16), "r" (val) );
+	out = (int16_t) (tmp & 0xffff); // not sure if the & 0xffff is necessary. test.
+	return out;
+}
+
 // computes ((a[31:0] * b[15:0]) >> 16)
 static inline int32_t signed_multiply_32x16b(int32_t a, uint32_t b) __attribute__((always_inline, unused));
 static inline int32_t signed_multiply_32x16b(int32_t a, uint32_t b)
@@ -129,7 +140,7 @@ static inline uint32_t pack_16x16(int32_t a, int32_t b)
 	return out;
 }
 
-// computes (((a[31:16] + b[31:16]) << 16) | (a[15:0 + b[15:0]))
+// computes (((a[31:16] + b[31:16]) << 16) | (a[15:0 + b[15:0]))  (saturates)
 static inline uint32_t signed_add_16_and_16(uint32_t a, uint32_t b) __attribute__((always_inline, unused));
 static inline uint32_t signed_add_16_and_16(uint32_t a, uint32_t b)
 {
@@ -137,6 +148,36 @@ static inline uint32_t signed_add_16_and_16(uint32_t a, uint32_t b)
 	asm volatile("qadd16 %0, %1, %2" : "=r" (out) : "r" (a), "r" (b));
 	return out;
 }
+
+// computes (((a[31:16] - b[31:16]) << 16) | (a[15:0 - b[15:0]))  (saturates)
+static inline int32_t signed_subtract_16_and_16(int32_t a, int32_t b) __attribute__((always_inline, unused));
+static inline int32_t signed_subtract_16_and_16(int32_t a, int32_t b)
+{
+	int32_t out;
+	asm volatile("qsub16 %0, %1, %2" : "=r" (out) : "r" (a), "r" (b));
+	return out;
+}
+
+// computes out = (((a[31:16]+b[31:16])/2) <<16) | ((a[15:0]+b[15:0])/2)
+static inline int32_t signed_halving_add_16_and_16(int32_t a, int32_t b) __attribute__((always_inline, unused));
+static inline int32_t signed_halving_add_16_and_16(int32_t a, int32_t b)
+{
+	int32_t out;
+	asm volatile("shadd16 %0, %1, %2" : "=r" (out) : "r" (a), "r" (b));
+	return out;
+}
+
+// computes out = (((a[31:16]-b[31:16])/2) <<16) | ((a[15:0]-b[15:0])/2)
+static inline int32_t signed_halving_subtract_16_and_16(int32_t a, int32_t b) __attribute__((always_inline, unused));
+static inline int32_t signed_halving_subtract_16_and_16(int32_t a, int32_t b)
+{
+	int32_t out;
+	asm volatile("shsub16 %0, %1, %2" : "=r" (out) : "r" (a), "r" (b));
+	return out;
+}
+
+
+
 
 // computes (sum + ((a[31:0] * b[15:0]) >> 16))
 static inline int32_t signed_multiply_accumulate_32x16b(int32_t sum, int32_t a, uint32_t b) __attribute__((always_inline, unused));
